(struct re_dfa_t): Change type of nodes_alloc and nodes_len to size_t.

+++ posix/regex_internal.h
@@ -625,8 +625,8 @@ struct re_fail_stack_t
 struct re_dfa_t
 {
   re_token_t *nodes;
-  int nodes_alloc;
-  int nodes_len;
+  size_t nodes_alloc;
+  size_t nodes_len;
   int *nexts;
   int *org_indices;
   re_node_set *edests;
